Refactor: factor out account contract
